Refactor object browser components and enhance directory listing
- Replaced `ObjectBrowser` with `DirectoryList`, `ObjectSummary`, and `ObjectPreview` components for improved structure and functionality.
- Updated `page.tsx` to conditionally render file details and previews based on object type.
- Introduced `DirectoryRow` for rendering individual items in the directory list.
- Removed deprecated `ObjectBrowser` and `ObjectDetails` components to streamline the codebase.
- Enhanced the directory listing logic to sort and display items more effectively.

This refactor aims to improve maintainability and user experience when navigating product files.
