+++ src/components/features/products/index.ts
@@ -1,13 +1,13 @@
-export { ProductListItem } from "./ProductListItem";
-export { ProductSummaryCard } from "./ProductSummaryCard";
-export { ProductMetaContent } from "./ProductMetaContent";
-export { ProductMetaCard } from "./ProductMetaCard";
-export { ProductHeader } from "./ProductHeader";
-export { ProductSchemaMetadata } from "./ProductSchemaMetadata";
-export { ProductSearchResult } from "./ProductSearchResult";
-export { TagList } from "./TagList";
-export { ObjectBrowser } from "./object-browser/ObjectBrowser";
-export { ChecksumVerifier } from "./ChecksumVerifier";
-export { ProductsFilters } from "./ProductsFilters";
-export { ProductsList } from "./ProductsList";
-export { ProductsSkeleton } from "./ProductsSkeleton";
+export * from "./ProductListItem";
+export * from "./ProductSummaryCard";
+export * from "./ProductMetaContent";
+export * from "./ProductMetaCard";
+export * from "./ProductHeader";
+export * from "./ProductSchemaMetadata";
+export * from "./ProductSearchResult";
+export * from "./TagList";
+export * from "./object-browser";
+export * from "./ChecksumVerifier";
+export * from "./ProductsFilters";
+export * from "./ProductsList";
+export * from "./ProductsSkeleton";
