+++ src/components/features/products/object-browser/index.ts
@@ -1,0 +1,3 @@
+export { DirectoryList } from "./DirectoryList";
+export { ObjectPreview } from "./ObjectPreview";
+export { ObjectSummary } from "./ObjectSummary";
