+++ src/components/index.ts
@@ -1,30 +1,22 @@
 // Core UI components
-export * from './core';
-export * from './display';
-export * from './layout';
+export * from "./core";
+export * from "./display";
+export * from "./layout";
 
 // Feature components
-export * from './features/markdown';
-export * from './features/metadata';
-export * from './features/profiles';
-export * from './features/products';
+export * from "./features/markdown";
+export * from "./features/metadata";
 export * from "./features/profiles";
-export * from './features/onboarding';
-export * from './features/keyboard';
+export * from "./features/products";
+export * from "./features/profiles";
+export * from "./features/onboarding";
+export * from "./features/keyboard";
 
 // Metadata components
-export { DataCiteSection, STACSection, ProductSchema, generateProductMetadata } from './features/metadata';
+export * from "./features/metadata";
 
 // Profile components
-export { AccountList, IndividualProfile, OrganizationProfile, ProfileLayout } from './features/profiles';
+export * from "./features/profiles";
 
 // Repository components
-export { 
-  ProductHeader,
-  ProductListItem,
-  ProductMetaCard,
-  ProductSchemaMetadata,
-  ProductSearchResult,
-  ProductSummaryCard,
-  ObjectBrowser
-} from './features/products'; 
+export * from "./features/products";
